Renamed the integration folder and files
